Refactored the solution to part1 to make part2 easier
I altered the code that converted a memory square to a coordinate to split out the part that calculates the coordinates in a spiral. I used a generator function so that it'll just keep calculating coordinates as long as you want, and then wrote a separate function that calls that to do the conversion from the memory square.