Adding a solution for day 5 part 2
I refactored out the code that executed each jump into a separate function so that I could inject a different implementation for part 2. Initially I made the functions return new copies of the state instead of mutating it, but then I realised that was causing the tests to take over 100 seconds to run. Changing the functions to just mutate the state got the times down to around 15 seconds, so I decided it made sense to do that.

I also removed the runInstructions.js file and the toString() function because I figured they weren't adding anything of value.