Adding a solution for day 6 part 1
I made a bit of a mess of this one and initially ended up with a solution that worked for the example in the puzzle description, but not the puzzle input. The problem was that I wasn't properly choosing the lowest index in the event of a tie when finding the highest bank.

I had actually written a test for that case, but I'd obviously made a mistake in it so it wasn't working quite right.