Adding a solution for day 7 part 2
I split the solution into a three stage process: one part builds the tree from the input; the second calculates the weight at each node in the tower; and the final stage finds the unbalanced node in the tree.

I added the outputTree.js script to output the tree to the console because initially I'd made a slight mistake and although I was finding the correct node, I wasn't calculating the new node correctly. This allowed me to take a look at it visually and figure out the problem.