Adding a solution for day 9 part 1
The solution parses the input into a tree structure and then walks the tree to calculate the score.

I ran into a bit of trouble where I wasn't parsing the garbage correctly, and was jumping an extra character when encountering a `!`. This is why I added the test case to try out the first set of garbage from the puzzle input.